create Middleware

+++ routes/web.php
@@ -20,11 +20,11 @@ Route::get('/', function () {
 });
 
 Route::get('/tasks/index', [TaskController::class, 'index'])->name('tasks.index');
-Route::get('/tasks/create', [TaskController::class, 'create'])->name('tasks.create');
+Route::get('/tasks/create', [TaskController::class, 'create'])->name('tasks.create')->middleware(['admin']);
 Route::get('/tasks/edit', [TaskController::class, 'edit'])->name('tasks.edit');
 Route::get('/tasks/{task}', [TaskController::class, 'show'])->name('tasks.show');
-Route::post('/tasks/', [TaskController::class, 'store'])->name('tasks.store');
-Route::put('/tasks/{task}', [TaskController::class, 'update'])->name('tasks.update');
-Route::delete('/tasks/{task}', [TaskController::class, 'delete'])->name('tasks.delete');
+Route::post('/tasks/', [TaskController::class, 'store'])->name('tasks.store')->middleware(['admin']);
+Route::put('/tasks/{task}', [TaskController::class, 'update'])->name('tasks.update')->middleware(['admin']);
+Route::delete('/tasks/{task}', [TaskController::class, 'delete'])->name('tasks.delete')->middleware(['admin']);
 
 Route::resource('/users', UserController::class);
